pyproject.toml
[project]
name = "youtube-video-transcript-ai-summary"
version = "0.1.0"
description = "CLI tool to generate AI summaries of YouTube video transcripts using Google's Gemini"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.2.0",
    "python-dotenv>=1.0.1",
    "typer>=0.15.1",
    "youtube-transcript-api>=0.6.3",
    "uv>=0.1.0",
    "pydantic>=2.10.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src"]

[project.scripts]
summarize = "src.main:summarize"
